LPS-64398 Web Facet implems: portal-search-facet to portal-search-web
https://issues.liferay.com/browse/LPS-64398
