+++ modules/apps/foundation/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/facet/FolderSearchFacet.java
@@ -12,26 +12,17 @@
  * details.
  */
 
-package com.liferay.portal.search.web.facet;
+package com.liferay.portal.search.web.internal.facet;
 
-import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
-import com.liferay.asset.kernel.model.AssetRendererFactory;
-import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.facet.AssetEntriesFacet;
+import com.liferay.portal.kernel.search.facet.FacetManager;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.search.facet.BaseJSPSearchFacet;
-import com.liferay.portal.search.facet.SearchFacet;
-
-import java.util.ArrayList;
-import java.util.List;
+import com.liferay.portal.search.facet.FacetConstants;
+import com.liferay.portal.search.web.facet.BaseJSPSearchFacet;
+import com.liferay.portal.search.web.facet.SearchFacet;
 
 import javax.portlet.ActionRequest;
 
@@ -44,18 +35,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = SearchFacet.class)
-public class AssetEntriesSearchFacet extends BaseJSPSearchFacet {
-
-	public List<AssetRendererFactory<?>> getAssetRendererFactories(
-		long companyId) {
-
-		return AssetRendererFactoryRegistryUtil.getAssetRendererFactories(
-			companyId);
-	}
+public class FolderSearchFacet extends BaseJSPSearchFacet {
 
 	@Override
 	public String getConfigurationJspPath() {
-		return "/facets/configuration/asset_entries.jsp";
+		return "/facets/configuration/folders.jsp";
 	}
 
 	@Override
@@ -67,14 +51,8 @@ public class AssetEntriesSearchFacet extends BaseJSPSearchFacet {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
 
 		jsonObject.put("frequencyThreshold", 1);
-
-		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
-
-		for (String assetType : getAssetTypes(companyId)) {
-			jsonArray.put(assetType);
-		}
-
-		jsonObject.put("values", jsonArray);
+		jsonObject.put("maxTerms", 10);
+		jsonObject.put("showAssetCount", true);
 
 		facetConfiguration.setDataJSONObject(jsonObject);
 
@@ -82,64 +60,52 @@ public class AssetEntriesSearchFacet extends BaseJSPSearchFacet {
 		facetConfiguration.setLabel(getLabel());
 		facetConfiguration.setOrder(getOrder());
 		facetConfiguration.setStatic(false);
-		facetConfiguration.setWeight(1.5);
+		facetConfiguration.setWeight(1.2);
 
 		return facetConfiguration;
 	}
 
 	@Override
 	public String getDisplayJspPath() {
-		return "/facets/view/asset_entries.jsp";
+		return "/facets/view/folders.jsp";
 	}
 
 	@Override
 	public String getFacetClassName() {
-		return AssetEntriesFacet.class.getName();
+		return FacetConstants.MULTI_VALUE_FACET;
 	}
 
 	@Override
 	public String getFieldName() {
-		return Field.ENTRY_CLASS_NAME;
+		return Field.FOLDER_ID;
 	}
 
 	@Override
 	public JSONObject getJSONData(ActionRequest actionRequest) {
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
 
 		int frequencyThreshold = ParamUtil.getInteger(
 			actionRequest, getClassName() + "frequencyThreshold", 1);
+		int maxTerms = ParamUtil.getInteger(
+			actionRequest, getClassName() + "maxTerms", 10);
+		boolean showAssetCount = ParamUtil.getBoolean(
+			actionRequest, getClassName() + "showAssetCount", true);
 
 		jsonObject.put("frequencyThreshold", frequencyThreshold);
-
-		String[] assetTypes = StringUtil.split(
-			ParamUtil.getString(actionRequest, getClassName() + "assetTypes"));
-
-		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
-
-		if (ArrayUtil.isEmpty(assetTypes)) {
-			assetTypes = getAssetTypes(themeDisplay.getCompanyId());
-		}
-
-		for (String assetType : assetTypes) {
-			jsonArray.put(assetType);
-		}
-
-		jsonObject.put("values", jsonArray);
+		jsonObject.put("maxTerms", maxTerms);
+		jsonObject.put("showAssetCount", showAssetCount);
 
 		return jsonObject;
 	}
 
 	@Override
 	public String getLabel() {
-		return "any-asset";
+		return "any-folder";
 	}
 
 	@Override
 	public String getTitle() {
-		return "asset-type";
+		return "folder";
 	}
 
 	@Override
@@ -150,25 +116,10 @@ public class AssetEntriesSearchFacet extends BaseJSPSearchFacet {
 		super.setServletContext(servletContext);
 	}
 
-	protected String[] getAssetTypes(long companyId) {
-		List<String> assetTypes = new ArrayList<>();
-
-		List<AssetRendererFactory<?>> assetRendererFactories =
-			AssetRendererFactoryRegistryUtil.getAssetRendererFactories(
-				companyId);
-
-		for (int i = 0; i < assetRendererFactories.size(); i++) {
-			AssetRendererFactory<?> assetRendererFactory =
-				assetRendererFactories.get(i);
-
-			if (!assetRendererFactory.isSearchable()) {
-				continue;
-			}
-
-			assetTypes.add(assetRendererFactory.getClassName());
-		}
-
-		return ArrayUtil.toStringArray(assetTypes);
+	@Override
+	@Reference(unbind = "-")
+	protected void setFacetManager(FacetManager facetManager) {
+		super.setFacetManager(facetManager);
 	}
 
 }
