+++ modules/apps/foundation/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/facet/ModifiedSearchFacet.java
@@ -12,16 +12,19 @@
  * details.
  */
 
-package com.liferay.portal.search.web.facet;
+package com.liferay.portal.search.web.internal.facet;
 
+import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.facet.MultiValueFacet;
+import com.liferay.portal.kernel.search.facet.FacetManager;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.search.facet.BaseJSPSearchFacet;
-import com.liferay.portal.search.facet.SearchFacet;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.search.facet.FacetConstants;
+import com.liferay.portal.search.web.facet.BaseJSPSearchFacet;
+import com.liferay.portal.search.web.facet.SearchFacet;
 
 import javax.portlet.ActionRequest;
 
@@ -34,11 +37,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = SearchFacet.class)
-public class AssetCategoriesSearchFacet extends BaseJSPSearchFacet {
+public class ModifiedSearchFacet extends BaseJSPSearchFacet {
 
 	@Override
 	public String getConfigurationJspPath() {
-		return "/facets/configuration/asset_categories.jsp";
+		return "/facets/configuration/modified.jsp";
 	}
 
 	@Override
@@ -49,66 +52,88 @@ public class AssetCategoriesSearchFacet extends BaseJSPSearchFacet {
 
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
 
-		jsonObject.put("displayStyle", "list");
-		jsonObject.put("frequencyThreshold", 1);
-		jsonObject.put("maxTerms", 10);
-		jsonObject.put("showAssetCount", true);
+		jsonObject.put("frequencyThreshold", 0);
+
+		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
+
+		for (int i = 0; i < _LABELS.length; i++) {
+			JSONObject range = JSONFactoryUtil.createJSONObject();
+
+			range.put("label", _LABELS[i]);
+			range.put("range", _RANGES[i]);
+
+			jsonArray.put(range);
+		}
+
+		jsonObject.put("ranges", jsonArray);
 
 		facetConfiguration.setDataJSONObject(jsonObject);
-
 		facetConfiguration.setFieldName(getFieldName());
 		facetConfiguration.setLabel(getLabel());
 		facetConfiguration.setOrder(getOrder());
 		facetConfiguration.setStatic(false);
-		facetConfiguration.setWeight(1.3);
+		facetConfiguration.setWeight(1.0);
 
 		return facetConfiguration;
 	}
 
 	@Override
 	public String getDisplayJspPath() {
-		return "/facets/view/asset_categories.jsp";
+		return "/facets/view/modified.jsp";
 	}
 
 	@Override
 	public String getFacetClassName() {
-		return MultiValueFacet.class.getName();
+		return FacetConstants.MODIFIED_FACET;
 	}
 
 	@Override
 	public String getFieldName() {
-		return Field.ASSET_CATEGORY_IDS;
+		return Field.MODIFIED_DATE;
 	}
 
 	@Override
 	public JSONObject getJSONData(ActionRequest actionRequest) {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
 
-		String displayStyleFacet = ParamUtil.getString(
-			actionRequest, getClassName() + "displayStyleFacet", "list");
 		int frequencyThreshold = ParamUtil.getInteger(
 			actionRequest, getClassName() + "frequencyThreshold", 1);
-		int maxTerms = ParamUtil.getInteger(
-			actionRequest, getClassName() + "maxTerms", 10);
-		boolean showAssetCount = ParamUtil.getBoolean(
-			actionRequest, getClassName() + "showAssetCount", true);
 
-		jsonObject.put("displayStyle", displayStyleFacet);
 		jsonObject.put("frequencyThreshold", frequencyThreshold);
-		jsonObject.put("maxTerms", maxTerms);
-		jsonObject.put("showAssetCount", showAssetCount);
+
+		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
+
+		String[] rangesIndexes = StringUtil.split(
+			ParamUtil.getString(
+				actionRequest, getClassName() + "rangesIndexes"));
+
+		for (String rangesIndex : rangesIndexes) {
+			JSONObject rangeJSONObject = JSONFactoryUtil.createJSONObject();
+
+			String label = ParamUtil.getString(
+				actionRequest, getClassName() + "label_" + rangesIndex);
+			String range = ParamUtil.getString(
+				actionRequest, getClassName() + "range_" + rangesIndex);
+
+			rangeJSONObject.put("label", label);
+			rangeJSONObject.put("range", range);
+
+			jsonArray.put(rangeJSONObject);
+		}
+
+		jsonObject.put("ranges", jsonArray);
 
 		return jsonObject;
 	}
 
 	@Override
 	public String getLabel() {
-		return "any-category";
+		return "any-time";
 	}
 
 	@Override
 	public String getTitle() {
-		return "category";
+		return "modified-date";
 	}
 
 	@Override
@@ -118,5 +143,20 @@ public class AssetCategoriesSearchFacet extends BaseJSPSearchFacet {
 	public void setServletContext(ServletContext servletContext) {
 		super.setServletContext(servletContext);
 	}
+
+	@Override
+	@Reference(unbind = "-")
+	protected void setFacetManager(FacetManager facetManager) {
+		super.setFacetManager(facetManager);
+	}
+
+	private static final String[] _LABELS = new String[] {
+		"past-hour", "past-24-hours", "past-week", "past-month", "past-year"
+	};
+
+	private static final String[] _RANGES = new String[] {
+		"[past-hour TO *]", "[past-24-hours TO *]", "[past-week TO *]",
+		"[past-month TO *]", "[past-year TO *]"
+	};
 
 }
